module-3/streaming_dataset/requirements.txt
# Core dependencies
mosaicml-streaming>=0.6.0
numpy>=1.20.0
pandas>=1.3.0
torch>=1.10.0
pillow>=8.0.0
tqdm>=4.62.0

# For benchmarking
matplotlib>=3.4.0
tabulate>=0.8.9
psutil>=5.8.0

# For MinIO integration (optional)
minio>=7.1.0
